% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{g}
\alias{g}
\title{normalize a vector}
\usage{
g(v)
}
\arguments{
\item{v.}{Numeric vector.}
}
\value{
Returns a vector of same length as \code{v} formed by dividing \code{v} with its \code{l2-norm}.
}
\description{
normalize a vector
}
