% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_l_testing.R
\name{l.ci_adjusted}
\alias{l.ci_adjusted}
\title{computing the post-LASSO-selection-adjusted ell-confidence interval}
\usage{
l.ci_adjusted(
  y,
  X,
  ind,
  gamma_range,
  lambda,
  lambda_cv = -1,
  coverage = 0.95,
  display = FALSE,
  smoothed = FALSE,
  outer_approx = FALSE,
  outer_grid.length = 10
)
}
\arguments{
\item{y}{Numeric vector. Response variable.}

\item{X}{Matrix or data frame. Full column-rank \verb{n x p} design matrix; \code{nrow(X)} must match \code{length(y)}.
The intercept is included.}

\item{ind}{Integer. Index of the coefficient to test.}

\item{lambda}{Numeric. Regularization parameter for the selection LASSO.}

\item{lambda_cv}{Numeric. Regularization parameter for evaluating the ell-distribution.
Either a positive value or one of:
\itemize{
\item \code{-1} to choose \code{lambda.min} with \code{cv.glmnet()} called on \verb{(y, X[,-ind])},
\item \code{-2} to choose \code{lambda.1se} with \code{cv.glmnet()} called on \verb{(y, X[,-ind])}. Default \code{-1}.
}}

\item{coverage}{the coverage of the confidence interval. Default is 0.95.}
}
\value{
returns a two-dimensional vector specifying the upper and lower limits of the ell-test confidence interval (adjusted for LASSO selection). If LASSO does not select variable \code{ind}, returns \code{NA}.
}
\description{
computing the post-LASSO-selection-adjusted ell-confidence interval
}
\examples{
set.seed(1)

n <- 100
p <- 50
s <- 5
A <- 2.3

X <- matrix(rnorm(n * p), nrow = n)
X <- apply(X, 2, g)  # normalize columns; recommended

beta <- rep(0, p)
rand_ind <- sample(1:p, size = s, replace = FALSE)
j <- rand_ind[1]     # index to test
beta[rand_ind] <- (1 - 2 * rbinom(s, 1, 0.5)) * A

y <- as.numeric(X \%*\% beta + rnorm(n))

gamma_range = seq(from = beta[j]-10, to = beta[j]+10, length.out = 100) 
#the grid of \gamma values to test on
ci_l_adjusted = l.ci_adjusted(y,X,j, gamma_range = gamma_range, coverage = 0.95, lambda = 0.01) 
#post-selection l-CI for \beta_j valid conditionally on LASSO with penalty 0.01 
#selecting the coefficient
}
