% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{qhaar}
\alias{qhaar}
\title{tail probabilities of the distribution of u1}
\usage{
qhaar(
  q,
  n,
  lower.tail = TRUE,
  stoperr = FALSE,
  known_sigma = FALSE,
  sigma.hat = 1
)
}
\arguments{
\item{q}{Numeric. \code{u1} value.}

\item{n}{Integer. The degrees of freedom; implies that \code{u} lies in an \code{n-1}-dimensional sphere.}

\item{lower.tail.}{Logical. If \code{TRUE}, returns the lower tail probability; if \code{FALSE}, returns the upper tail probability. Default is \code{TRUE}.}

\item{`known_sigma`.}{Logical. If \code{TRUE}, returns probabilities of a normal distribution with standard deviation \code{sigma.hat} (mentioned below) instead of the \code{u1}-distribution.
Default is FALSE.}

\item{sigma.hat.}{Numeric. If \code{known_sigma = TRUE}, this is the standard deviation of the normal distribution. Default is 1.}
}
\value{
Returns the percentile of \code{q} with respect to the \code{u1}-distribution or the normal distribution corresponding to whether
\code{known_sigma} is \code{FALSE} or \code{TRUE}.
}
\description{
tail probabilities of the distribution of u1
}
