% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{std}
\alias{std}
\title{standardize a vector}
\usage{
std(y)
}
\arguments{
\item{y.}{Numeric vector.}
}
\value{
Returns vector of same length as \code{y} formed by standardizing \code{y}.
}
\description{
standardize a vector
}
